Major refactoring first step: Create libraries.
