+++ lib/CoffeeAutomation/GrindingController.cpp
@@ -1,0 +1,123 @@
+#include <Constants.hpp>
+#include <GrindingController.hpp>
+
+//----------- In/Out -----------
+constexpr int relay = LED_BUILTIN;
+
+//----------- Scale -----------
+constexpr int scaleData = 4;
+constexpr int scaleClk = 5;
+constexpr float scaleCalibrationFactor = 418;
+
+GrindingController::GrindingController() : mqttGrinder() {}
+
+void GrindingController::setGrindingTime(unsigned int timeToGrind) {
+  grindingTime = timeToGrind;
+  mqttGrinder.publishMqttTopicAndValue(topicOutCurrentSetGrindingTime,
+                                       grindingTime);
+}
+
+void GrindingController::startGrinding(const char* startTriggerOrigin,
+                                       unsigned int timeToGrind) {
+  grindingOngoing = true;
+  grindingStartedTime = millis();
+  setGrindingTime(timeToGrind);
+  mqttGrinder.publishMqttTopicAndValue(topicOutStarted, startTriggerOrigin);
+  digitalWrite(relay, HIGH);
+}
+
+float GrindingController::getCurrentWeightAndPublish() {
+  auto currentWeight = scale.get_units(1);
+  if ((millis() - lastScaleUpdate) > scaleUpdateTime) {
+    lastScaleUpdate = millis();
+    mqttGrinder.publishMqttTopicAndValue(topicOutCurrentWeight, currentWeight);
+  }
+  return currentWeight;
+}
+
+unsigned int getIntFromPayload(byte* payload, unsigned int length) {
+  payload[length] = '\0';
+  String s = String((char*)payload);
+  return s.toInt();
+}
+
+void GrindingController::callback(char* topic,
+                                  byte* payload,
+                                  unsigned int length) {
+  Serial.print("Received message [");
+  Serial.print(topic);
+  Serial.println("] ");
+
+  if (strcmp(topic, topicInSetGrindingTime) == 0) {
+    setGrindingTime(getIntFromPayload(payload, length));
+    Serial.print("-> ");
+    Serial.println(grindingTime);
+  } else if (strcmp(topic, topicInStart) == 0) {
+    if (length > 0) {
+      startGrinding("callbackWithTime", getIntFromPayload(payload, length));
+    } else {
+      startGrinding("callbackWithoutTime", grindingTime);
+    }
+  } else if (strcmp(topic, topicInTare) == 0) {
+    scale.tare(5);
+  } else if (strcmp(topic, topicInAutomatic) == 0) {
+    scale.tare(5);
+    if (length > 0) {
+      desiredGrams = ((float)getIntFromPayload(payload, length) - 0.2);
+    } else {
+      desiredGrams = 16.8;
+    }
+    getCurrentWeightAndPublish();
+    automaticGrindingOngoing = true;
+  } else {
+    Serial.print("Unhandled topic '");
+    Serial.print(topic);
+    Serial.println("'!");
+  }
+}
+
+void GrindingController::setup() {
+  pinMode(relay, OUTPUT);
+  digitalWrite(relay, LOW);
+  mqttGrinder.setup(std::bind(&GrindingController::callback, this,
+                              std::placeholders::_1, std::placeholders::_2,
+                              std::placeholders::_3));
+
+  scale.begin(scaleData, scaleClk);
+  scale.set_scale(scaleCalibrationFactor);
+  scale.tare();
+}
+
+void GrindingController::loop() {
+  mqttGrinder.loop();
+  getCurrentWeightAndPublish();
+
+  if (automaticGrindingOngoing) {
+    auto currentWeight = getCurrentWeightAndPublish();
+    if ((currentWeight + thresholdTargetGrams) < desiredGrams) {
+      mqttGrinder.publishMqttTopicAndValue(topicOutStarted,
+                                           "automaticGrinding");
+      digitalWrite(relay, HIGH);
+    } else {
+      digitalWrite(relay, LOW);
+      delay(500);
+      currentWeight = getCurrentWeightAndPublish();
+      if (currentWeight < desiredGrams) {
+        startGrinding("automaticGrinding", 150);
+      } else {
+        digitalWrite(relay, LOW);
+        automaticGrindingOngoing = false;
+        mqttGrinder.publishMqttTopicAndValue(topicOutAutomaticFinished,
+                                             currentWeight);
+      }
+    }
+
+  } else if (grindingOngoing &&
+             (millis() - grindingStartedTime) > grindingTime) {
+    grindingOngoing = false;
+    digitalWrite(relay, LOW);
+    mqttGrinder.publishMqttTopicAndValue(topicOutFinished, grindingTime);
+  } else {
+    // ArduinoOTA.handle();
+  }
+}
